+++ node/src/main/kotlin/net/corda/node/services/statemachine/transitions/StateMachine.kt
@@ -10,8 +10,9 @@
 
 package net.corda.node.services.statemachine.transitions
 
-import net.corda.core.flows.*
-import net.corda.node.services.statemachine.*
+import net.corda.core.flows.StateMachineRunId
+import net.corda.node.services.statemachine.Event
+import net.corda.node.services.statemachine.StateMachineState
 import java.security.SecureRandom
 
 /**
@@ -37,30 +38,11 @@ enum class SessionDeliverPersistenceStrategy {
     OnNextCommit
 }
 
-/**
- * @property sessionDeliverPersistenceStrategy see [SessionDeliverPersistenceStrategy]
- * @property eventQueueSize the size of a flow's event queue. If the queue gets full the thread scheduling the event
- *     will block. An example scenario would be if the flow is waiting for a lot of messages at once, but is slow at
- *     processing each.
- */
-data class StateMachineConfiguration(
-        val sessionDeliverPersistenceStrategy: SessionDeliverPersistenceStrategy,
-        val eventQueueSize: Int
-) {
-    companion object {
-        val default = StateMachineConfiguration(
-                sessionDeliverPersistenceStrategy = SessionDeliverPersistenceStrategy.OnDeliver,
-                eventQueueSize = 16
-        )
-    }
-}
-
 class StateMachine(
         val id: StateMachineRunId,
-        val configuration: StateMachineConfiguration,
         val secureRandom: SecureRandom
 ) {
     fun transition(event: Event, state: StateMachineState): TransitionResult {
-        return TopLevelTransition(TransitionContext(id, configuration, secureRandom), state, event).transition()
+        return TopLevelTransition(TransitionContext(id, secureRandom), state, event).transition()
     }
 }