+++ docs/source/changelog.rst
@@ -8,6 +8,8 @@ Unreleased
 ==========
 * Introduced a hierarchy of ``DatabaseMigrationException``s, allowing ``NodeStartup`` to gracefully inform users of problems related to database migrations before exiting with a non-zero code.
 
+* Fixed an issue with ``CashException`` not being able to deserialise after the introduction of AMQP for RPC.
+
 * Removed -xmx VM argument from Explorer's Capsule setup. This helps avoiding out of memory errors.
 
 * Shell now kills an ongoing flow when CTRL+C is pressed in the terminal.
@@ -54,7 +56,10 @@ Unreleased
     The encoded bytes are also serialised into the ``encoded`` field. This can be used to deserialise an ``X509Certificate``
     back.
   * ``CertPath`` objects are serialised as a list of ``X509Certificate`` objects.
-  * ``SignedTransaction`` is serialised into its ``txBits`` and ``signatures`` and can be deserialised back
+  * ``WireTransaction`` now nicely outputs into its components: ``id``, ``notary``, ``inputs``, ``attachments``, ``outputs``,
+    ``commands``, ``timeWindow`` and ``privacySalt``. This can be deserialised back.
+  * ``SignedTransaction`` is serialised into ``wire`` (i.e. currently only ``WireTransaction`` tested) and ``signatures``,
+    and can be deserialised back.
 
 * ``fullParties`` boolean parameter added to ``JacksonSupport.createDefaultMapper`` and ``createNonRpcMapper``. If ``true``
   then ``Party`` objects are serialised as JSON objects with the ``name`` and ``owningKey`` fields. For ``PartyAndCertificate``